// microservices/arca-ms/src/enum/iva-condition.enum.ts
export enum IvaCondition {
  RESPONSABLE_INSCRIPTO = 'RESPONSABLE_INSCRIPTO',
  MONOTRIBUTISTA = 'MONOTRIBUTISTA',
  EXENTO = 'EXENTO',
  CONSUMIDOR_FINAL = 'CONSUMIDOR_FINAL',
  SUJETO_NO_CATEGORIZADO = 'SUJETO_NO_CATEGORIZADO',
  PROVEEDOR_DEL_EXTERIOR = 'PROVEEDOR_DEL_EXTERIOR',
  CLIENTE_DEL_EXTERIOR = 'CLIENTE_DEL_EXTERIOR',
  IVA_LIBERADO_LEY_19640 = 'IVA_LIBERADO_LEY_19640',
  IVA_NO_ALCANZADO = 'IVA_NO_ALCANZADO',
}